test: update wallet related tests
