+++ src/app/wallets/wallets-actions/wallets-actions.component.html
@@ -2,7 +2,7 @@
 	<ion-button
 		color="primary-light"
 		(click)="generateWalletHandler()"
-		data-testid="wallets-actions-__button--generate"
+		data-testid="wallets-actions__button--generate"
 	>
 		<ion-icon size="small" slot="icon-only" name="add-sharp"></ion-icon>
 	</ion-button>
